cleaning and refactoring part 2

+++ src/main/java/com/github/pinmacaroon/dchook/util/TimeConverter.java
@@ -2,6 +2,6 @@ package com.github.pinmacaroon.dchook.util;
 
 public class TimeConverter {
     public static String timeOfDayToHoursMinutes(long time){
-        return (int) Math.floor(time / 1000) + ":" + (int) ((time % 1000) / 1000.0 * 60);
+        return (int) Math.floor((double) time / 1000) + ":" + (int) ((time % 1000) / 1000.0 * 60);
     }
 }
